+++ packages/bundler/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add health check to server
 - fix bug with marking op reverted and NFs unused (redis txs misformatted)
 - bump sdk with joinsplit sorting and note timestamp changes
 - add `gasAssetRefundThreshold` to operation for req validation