+++ packages/offchain-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add `HealthCheckResponse`
 - move redis, ajv, and actor related utils to this pkg so bundler and screener can share code
 - `makeLogger` adds console exceptionHandlers/rejectionHandlers if `consoleLevel` is defined
 - add loggers